move Fourier to utils
