+++ docs/source/transforms.rst
@@ -53,11 +53,6 @@ Generic Interfaces
 .. autoclass:: Decollated
     :members:
 
-`Fourier`
-^^^^^^^^^^^^^
-.. autoclass:: Fourier
-    :members:
-
 Vanilla Transforms
 ------------------
