+++ monai/transforms/intensity/array.py
@@ -24,8 +24,8 @@ from monai.config import DtypeLike
 from monai.config.type_definitions import NdarrayTensor
 from monai.data.utils import get_random_patch, get_valid_patch_size
 from monai.networks.layers import GaussianFilter, HilbertTransform, SavitzkyGolayFilter
-from monai.transforms.transform import Fourier, RandomizableTransform, Transform
-from monai.transforms.utils import equalize_hist, is_positive, rescale_array
+from monai.transforms.transform import RandomizableTransform, Transform
+from monai.transforms.utils import Fourier, equalize_hist, is_positive, rescale_array
 from monai.utils import (
     PT_BEFORE_1_7,
     InvalidPyTorchVersionError,
